refactor: split up types
