+++ libs/core/src/lib/validations/empty/empty.ts
@@ -1,5 +1,5 @@
 import { LengthProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 
 /**
