+++ libs/core/src/lib/validations/greater-than-or-equals/greater-than-or-equals.ts
@@ -1,5 +1,5 @@
 import { NumberProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 import { DEFAULT_PLACEHOLDERS } from '../message-formatter';
