+++ libs/core/src/lib/validations/is-true/is-true.ts
@@ -1,5 +1,5 @@
 import { BooleanProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 
 /**
