+++ libs/core/src/lib/validations/matches/matches.ts
@@ -1,5 +1,5 @@
 import { StringProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 
 /**
