+++ libs/core/src/lib/validations/max-length/max-length.ts
@@ -1,5 +1,5 @@
 import { LengthProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 import { DEFAULT_PLACEHOLDERS } from '../message-formatter';
