+++ libs/core/src/lib/validations/must/must.ts
@@ -1,4 +1,4 @@
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 
 /**
