+++ libs/core/src/lib/validations/not-empty/not-empty.ts
@@ -1,5 +1,5 @@
 import { LengthProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { empty } from '../empty/empty';
 import { not } from '../not/not';
