+++ libs/core/src/lib/validations/required/required.ts
@@ -1,5 +1,5 @@
 import { isLengthProperty } from '../../types/properties';
-import { SyncValidation } from '../../types/types';
+import { SyncValidation } from '../../types/validations';
 import { createValidation } from '../create-validation';
 
 /**
